% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppsAcoplaTaxaLivreRisco.R
\name{rppsAcoplaTaxaLivreRisco}
\alias{rppsAcoplaTaxaLivreRisco}
\title{Acopla uma taxa livre de risco a serie temporal de retornos dos investimentos em cotas de fundos de um RPPS.}
\usage{
rppsAcoplaTaxaLivreRisco(retornos, taxa, nome.taxa)
}
\arguments{
\item{retornos}{objeto xts contendo os retornos mensais da carteira de investimentos em fundos do RPPS, conforme
gerado pela funcao \code{rppsDesempenho}}

\item{taxa}{data.frame contendo a serie historica da taxa livre de risco brasileira. Assume-se que
o data.frame possui as seguintes colunas DATA e VALOR, nesta ordem.}

\item{nome.taxa}{nome do titulo a ser utilizado como taxa livre de risco.}
}
\value{
um objeto do tipo data.frame contendo os retornos do portfolio das aplicacoes financeiras do RPPS, acrescidos
        da taxa livre de risco da economia
}
\description{
Esta funcao acopla uma taxa livre de risco a serie temporal de retornos da carteiras de investimentos
em cotas de fundos obtida a partir da funcao \code{rppsDesempenho}.
}
\details{
Obs.: A serie temporal de retornos so podera conter uma taxa livre de risco. Se chamada mais de uma
      vez para uma mesma serie historica de retornos, nova taxa sendo incluida ira substituir outra
      que esteja presente.
}
\examples{
\dontrun{
tsRetornos <- rppsDesempenho(dtCvm, dtDair)
tsRetornos <- rppsAcoplaTaxaLivreRisco(retornos = tsRetornos,
                                       taxa = dfLft2021,
                                       nome.taxa = "LFT 2021")
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\code{RppsWrangler} e \code{RcvmWrangler}
}

