% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppsDesempenho.R
\name{rppsDesempenho}
\alias{rppsDesempenho}
\title{Calculo do desempenho da carteira de investimentos em fundos de um RPPS.}
\usage{
rppsDesempenho(dtCvm, dtDair)
}
\arguments{
\item{dtCvm}{data.frame contendo dados extraidos dos informes diarios de fundos da CVM, com a mesma
estrutura retornada pela funcao de extracao de informes do pacote \code{RcvmWrangler}.}

\item{dtDair}{data.frame contendo dados extraidos do "DAIR", com a mesma estrutura retornada pela funcao
de extracao de DAIR do pacote \code{RppsWrangler}.}
}
\value{
um objeto do tipo data.frame contendo os retornos do portfolio das aplicacoes financeiras do RPPS.
}
\description{
Esta funcao calcula, a partir dos dados contidos no Demonstrativo de Aplicacoes e Investimentos dos Recursos (DAIR)
dos RPPS e dos informes diarios dos fundos de investimentos disponibilizados pela CVM, o retorno da carteira de
investimentos em fundos de um RPPS.
}
\details{
Obs.: O desempenho calculado esta aproximado pois as informacoes da composicao da carteira do RPPS so
sao informados com periodicidade bimestral.
}
\examples{
\dontrun{
tsRetornos <- rppsDesempenho(dtCvm, dtDair)
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\code{RppsWrangler} e \code{RcvmWrangler}
}

