% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppsPlotaDesempenho.R
\name{rppsPlotaDesempenho}
\alias{rppsPlotaDesempenho}
\title{Plota o desempenho da carteira de investimentos em fundos de um RPPS.}
\usage{
rppsPlotaDesempenho(rtRpps)
}
\arguments{
\item{rtRpps}{um objeto do tipo xts_dair contendo os retornos do portfolio das aplicacoes financeiras do RPPS.}
}
\value{
um objeto do tipo gtable contendo um grafico produzido com o pacote ggplot2
}
\description{
Esta funcao recebe um objeto do tipo xts_dair, contendo o retorno da carteira de um RPPS, bem como
dos retornos dos fundos que a compoem, e retorna um grafico do tipo ggplot por meio de um
objeto gtable
}
\examples{
\dontrun{
tsRetornos <- rppsDesempenho(dtCvm, dtDair)
rppsPlotaDesempenho(tsRetornos)
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\code{PerformanceAnalytics}, \code{xts}, \code{ggplot2}, \code{gridExtra}
}

